Fixed #8298: added ZStream.readFile support for ScalaNative
